requirements.txt
Django==1.10
argparse==1.2.1
django-crispy-forms==1.6.0
django-registration-redux==1.4
wsgiref==0.1.2
